with some graphics

+++ srcs/data_carct.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   data_carct.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aaires-b <aaires-b@@student.42.fr>         +#+  +:+       +#+        */
+/*   By: aaires-b <aaires-b@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/15 15:44:18 by aaires-b          #+#    #+#             */
-/*   Updated: 2023/10/15 15:48:54 by aaires-b         ###   ########.fr       */
+/*   Updated: 2023/10/18 17:17:33 by aaires-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int data_caract(char **newmap)
 {
-	int x;
-	int y;
+	float x;
+	float y;
 
 	x = 0;
 	while(newmap[x])
